Updated E - Area of a Circle.c changed doubles to floats

+++ ICPC Assiut University Community ( Newcomers Training Sheets )/Sheet #1 (Data type - Conditions)/E - Area of a Circle.c
@@ -2,10 +2,10 @@
 #include <math.h>
  
 int main() {
-    const double PI = 3.141592653;
-    double r, area;
-    scanf("%lf",&r);
+    const float PI = 3.141592653;
+    float r, area;
+    scanf("%f",&r);
     area = PI * r *r;
-    printf("%.9lf", area);
+    printf("%.9f", area);
     return 0;
 }
